models: fix review issues

+++ app/models/reseller_product.py
@@ -12,4 +12,4 @@ class ResellerProduct(db.Model):
     product_id = db.Column(db.Integer, db.ForeignKey("product.id"))
     reseller_id = db.Column(db.Integer, db.ForeignKey("reseller.id"))
     months = db.Column(db.Integer)
-    price = db.Column(db.Integer)
+    price = db.Column(db.Float)
